requirements.txt
pyexiv2
python-dateutil
